requirements.txt
discord.py
bs4
requests
PyNaCl